renamed UserModel to UserSettingsModel and changed its structure
